Fetch Address from Explorer.Indexer.BlockFetcher
It removes the responsability to provide a ready-to-validate set
of Addresses hashes from the Explorer.Chain context.

All Addresses that are created during the block indexing are
collected from previously Blocks, Transactions, InternalTransactions and
Logs fetched data. The Address data was built upon that in the Chain
context after running all the other entities validations and casting, which
was its major concern/responsability.

We've moved this Address data fetching out of the Chain context, giving
this responsability to the BlockFetcher (as it occurs with the other
entity). This allowed us to remove a few functions that were included
in all entities modules only to support the Address data fetching.