Allow pending transactions in Transaction.to_elixir
Pending transactions have nil for "blockNumber" and "blockHash", which
broke `quantity_to_integer`, so for those keys, check if `nil` and pass
through.  For others, check that the value is not `nil` as it gives
a better error message with both the key and value instead of just the
nil value passed to quantity_to_integer as happened before.