Add contract creator on Account details page
Now the user will be able to see the contract creator and the
transaction references to it on the Account details page when the
address is a contract.

In order to this work we had to make a query that returns the internal
transaction that has the contract creation. This way we can get the
`from_address` that represents the contract creator and the transaction
from the internal transaction.