Fix log in development environment
IMHO, it's very helpful to see which query is being generated by Ecto
and how long it's taking. For some reason it was disabled, I'm not sure
if it was intentional or a mistake.

If it was intentional, we can close this PR and keep just as reference
for developers that want to enabled these logs.