Don't log when queue is unavailable
Fixes #1223

For large chains like ETH Mainnet, it is fairly common for the catchup
fetcher to take a long time to calculate the missing block numbers and
start the named Sequence used by `push_front(block)`.  Since this is
normal and not exceptional, `InvalidConsensus.Worker` shouldn't log this
and retry silently.