fetch_and_import_range_from_sequence log exceptions instead of raise
Fixes #1234

Raising the exception will kill all Tasks, the sequence and the fetcher
because of the linking from `Task.async_stream`.  This leads to when 1
Task times out in a DB transaction, they all are affected, leading to
slowdowns when the sequence has to be rebuilt when the fetcher restarts.