Remove obsolete ConsensusEnsurer
As a result of #1657 out-of-place parent blocks are marked as non-consensus
more reliably, i.e. they will eventually be fetched even if first attempt is
unsuccessful.

The `ConsensusEnsurer` module, on the contrary, only tried refetching block
once, and left it as-is if the refetch failed. Now it is not needed anymore.