Implement logging in using MetaMask extension (#2408)
We check Web3 every 0.1s to check for currently selected account.
This is ugly and MetaMask developers admit that, but it's the only way they
can recommend.

Login button always redirects to metamask.io website, but if the MetaMask
is detected, it's onclick is overridden to launch MetaMask's Privacy Mode
unlock dialog.

When an account is detected, it is sent to websocket and used for rendering
top panel during further staking information updates.

Putting account into session is not implemented, as it _might_ cause privacy
issues. In the meanwhile, we'll try to think of more clever way for preventing
login button blinking during page load.

Co-Authored-By: Anatoly Nikiforov <anatolyniky@gmail.com>
Co-Authored-By: Kirill Andreev <hindmost.one@gmail.com>