Load contract ABIs after signing in, add makeContractCall utility
After user signs in using MetaMask for the first time, the contract addresses
and ABIs are pushed throught the websocket. They are saved in the state
for making calls to MetaMask API.

makeContractCall utility is added for that purpose: it makes a call, shows
success/error message from API, and shows error if 15s timeout passes.

Token information is passed as well for calculations, comparisons and alerts.