Separate ordered withdrawals from active stakes
Result of stakeAmount getter includes ordered withdrawals, which should not
actually be displayed as part of stake from user's standpoint.
Thus, we now use stakeAmountMinusOrderedWithdraw getter for that amount and
start showing the total amount of ordered withdrawals in top panel.

In addition, a couple of missing is_deleted filters is added.