Support ttl properly in oredered caches
Problem: the macro that creates ordered caches can receive ttl settings for the underlining ConCache process, but the index wont be updated.
This would cause the index's list to contain ids of elements that are actually not there anymore.

Solution: add a default callback function (overridable) that updates the index when necessary.