Add temporary indexer to trigger refetching of blocks with duplicated token transfers
Problem: There are inconsistencies in the database regarding blocks with duplicate token transfers.
A fix for new data is already mergeed and the first and last block affected are known, but they have to be corrected.

Solution: implement a temporary indexer that handles this range of blocks, triggering thir refetch.