Apply block filters before pagination in token transfers list query
Since the introduction of 21ca08b8994e56f1b329900e79d9927ee52d9351,
`start_block` and `end_block` filters are applied after pagination (via `page` and `offset`) is done.

The subquery will fetch `options.page_size` results, the main query will then do a join,
and the blocks filters will be applied on this query.
So assuming an address with more token transfers than `page_size`,
if the `start_block` is after the block of the last token transfer of the first page,
the query would return an empty result,
while setting `page` to a different number might result the result if hitting the correct page.

This changes the behavior to apply the blocks filters on the subquery.